refactor(guid): 将 index.module.css 移至 styles 子目录
