+++ src/renderer/pages/guid/components/AgentPillBar.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { getAgentLogo } from '@/renderer/utils/agentLogo';
 import { resolveExtensionAssetUrl } from '@/renderer/utils/platform';
 import type { AcpBackendAll } from '@/types/acpTypes';
-import styles from '../index.module.css';
+import styles from '../styles/index.module.css';
 import type { AvailableAgent } from '../types';
 
 type AgentPillBarProps = {
